--- a/purebred.cabal
+++ b/purebred.cabal
@@ -67,6 +67,8 @@ library
                      , Purebred.Types.Mailcap
                      , Purebred.Types.Parser.ByteString
                      , Purebred.Types.Parser.Text
+                     , Purebred.Types.Presentation
+                     , Purebred.Types.Presentation.MailBody
                      , Purebred.Types.UI
                      , Purebred.UI.Actions
                      , Purebred.UI.App
@@ -151,7 +153,6 @@ test-suite unit
   other-modules:       TestMail
                      , TestActions
                      , TestTagParser
-                     , TestTextParser
                      , LazyVector
   default-language:    Haskell2010
   build-depends:       base