--- a/src/Purebred/Config.hs
+++ b/src/Purebred/Config.hs
@@ -32,7 +32,6 @@ import qualified Data.Text as T
 import System.Directory (getHomeDirectory)
 
 import qualified Brick.AttrMap as A
-import qualified Brick.Widgets.List as L
 import qualified Brick.Widgets.Dialog as D
 import Brick.Util (fg, on, bg)
 import qualified Brick.Widgets.Edit as E
@@ -66,6 +65,7 @@ import qualified Purebred.Plugin.UserAgent
 import Purebred.System.Process
 import Purebred.Types.IFC (sanitiseText, untaint)
 import Purebred.Types.Error
+import Purebred.UI.Attr
 
 -- | Invoke a @sendmail(1)@-compatible program to send the email
 --
@@ -130,134 +130,6 @@ solarizedDark =
 -- These attributes are used as keys in widgets to assign color values.
 --
 
--- ** State Attributes used to indicate list item state
--- List attributes are generated based on three possible states:
---
---   * selected (first)
---   * toggled
---   * new (last)
---
--- That means that for example, any new list items will inherit colour
--- definitions from selected and toggled attributes.
-listStateSelectedAttr :: A.AttrName
-listStateSelectedAttr = A.attrName "selected"
-
-listStateNewmailAttr :: A.AttrName
-listStateNewmailAttr = A.attrName "newmail"
-
-listStateToggledAttr :: A.AttrName
-listStateToggledAttr = A.attrName "toggled"
-
--- ** Widget Attributes
---
-defaultAttr :: A.AttrName
-defaultAttr = A.attrName "default"
-
-mailViewAttr :: A.AttrName
-mailViewAttr = A.attrName "mailview"
-
-statusbarAttr :: A.AttrName
-statusbarAttr = A.attrName "statusbar"
-
-statusbarErrorAttr :: A.AttrName
-statusbarErrorAttr = statusbarAttr <> A.attrName "error"
-
-statusbarInfoAttr :: A.AttrName
-statusbarInfoAttr = statusbarAttr <> A.attrName "info"
-
-statusbarWarningAttr :: A.AttrName
-statusbarWarningAttr = statusbarAttr <> A.attrName "warning"
-
-editorAttr :: A.AttrName
-editorAttr = E.editAttr
-
-editorFocusedAttr :: A.AttrName
-editorFocusedAttr = E.editFocusedAttr
-
-editorErrorAttr :: A.AttrName
-editorErrorAttr = editorAttr <> A.attrName "error"
-
-editorLabelAttr :: A.AttrName
-editorLabelAttr = editorAttr <> A.attrName "label"
-
-listAttr :: A.AttrName
-listAttr = L.listAttr
-
--- Note: Brick exports a L.listSelectedAttr, yet in order to make our
--- use of our listState attributes consistent across the application
--- we need to use our own listState attributes.
-listSelectedAttr :: A.AttrName
-listSelectedAttr = L.listAttr <> listStateSelectedAttr
-
-listNewMailAttr :: A.AttrName
-listNewMailAttr = L.listAttr <> listStateNewmailAttr
-
-listSelectedNewmailAttr :: A.AttrName
-listSelectedNewmailAttr = L.listSelectedAttr <> listStateNewmailAttr
-
-listToggledAttr :: A.AttrName
-listToggledAttr = L.listAttr <> listStateToggledAttr
-
-listSelectedToggledAttr :: A.AttrName
-listSelectedToggledAttr = listStateSelectedAttr <> listToggledAttr
-
-mailAttr :: A.AttrName
-mailAttr = A.attrName "mail"
-
-mailTagAttr :: A.AttrName
-mailTagAttr = mailAttr <> A.attrName "tag"
-
-mailTagToggledAttr :: A.AttrName
-mailTagToggledAttr = mailTagAttr <> listStateToggledAttr
-
-mailAuthorsAttr :: A.AttrName
-mailAuthorsAttr = mailAttr <> A.attrName "authors"
-
-mailNewmailAuthorsAttr :: A.AttrName
-mailNewmailAuthorsAttr = mailAuthorsAttr <> listStateNewmailAttr
-
-mailToggledAuthorsAttr :: A.AttrName
-mailToggledAuthorsAttr = mailAuthorsAttr <> listStateToggledAttr
-
-mailSelectedAuthorsAttr :: A.AttrName
-mailSelectedAuthorsAttr = mailAuthorsAttr <> listStateSelectedAttr
-
-mailSelectedNewmailAuthorsAttr :: A.AttrName
-mailSelectedNewmailAuthorsAttr = mailAuthorsAttr <> listStateSelectedAttr <> listStateNewmailAttr
-
-mailSelectedToggledAuthorsAttr :: A.AttrName
-mailSelectedToggledAuthorsAttr = mailSelectedAuthorsAttr <> listStateToggledAttr
-
-headerAttr :: A.AttrName
-headerAttr = A.attrName "header"
-
-headerKeyAttr :: A.AttrName
-headerKeyAttr = headerAttr <> A.attrName "key"
-
-headerValueAttr :: A.AttrName
-headerValueAttr = headerAttr <> A.attrName "value"
-
-helpAttr :: A.AttrName
-helpAttr = A.attrName "help"
-
-helpTitleAttr :: A.AttrName
-helpTitleAttr = helpAttr <> A.attrName "title"
-
-helpKeybindingAttr :: A.AttrName
-helpKeybindingAttr = helpAttr <> A.attrName "keybinding"
-
-
-textMatchHighlightAttr :: A.AttrName
-textMatchHighlightAttr = A.attrName "match"
-
-currentTextMatchHighlightAttr :: A.AttrName
-currentTextMatchHighlightAttr = textMatchHighlightAttr <> A.attrName "current"
-
-mailbodyAttr :: A.AttrName
-mailbodyAttr = A.attrName "mailbody"
-
-mailbodySourceAttr :: A.AttrName
-mailbodySourceAttr = mailbodyAttr <> A.attrName "source"
 
 
 -- | Returns the notmuch database path by executing 'notmuch config