--- a/src/Purebred/UI/Help/Main.hs
+++ b/src/Purebred/UI/Help/Main.hs
@@ -35,8 +35,8 @@ import Graphics.Vty.Input.Events (Event(..), Key(..), Modifier(..))
 import Control.Lens (view, views, ifoldr)
 import qualified Data.Map.Strict as M
 import Data.Text (Text, singleton, intercalate, pack)
-import Purebred.Config (helpTitleAttr, helpKeybindingAttr)
 import Purebred.Types
+import Purebred.UI.Attr (helpTitleAttr, helpKeybindingAttr)
 import Purebred.UI.Utils (titleize, Titleize)
 
 data KeybindingHelp =