… src/Data/Text/Markup.hs
… src/Purebred/Brick/Markup.hs
… src/Purebred/UI/Actions.hs
… src/Purebred/UI/Attr.hs
… test/TestMail.hs
… test/TestTextParser.hs